requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
docker-py==1.10.6
docker-pycreds==0.4.0
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
visitor==0.1.3
websocket-client==0.57.0
Werkzeug==1.0.1
